% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/allbarplots.R
\name{allbarplots}
\alias{allbarplots}
\title{Create Barplots for all Rows in a Matrix}
\usage{
allbarplots(lookups, lists, fileroot = NULL, col = "red", height = 800,
  width = 800, res = 100, alpha = 50, ...)
}
\arguments{
\item{lookups}{matrix or named list of matrices,
each with symmetric row names and column names}

\item{lists}{character vector names from lookups}

\item{fileroot}{single character label to prepend to start of filenames,
perhaps including path, which will be constructed
"<fileroot><listname><lookupname>.png".
If NULL, plots will be created in the windows device.}

\item{col}{single colour for bars}

\item{height}{single integer, height of png image to create (px)}

\item{width}{single integer, width of png image to create (px)}

\item{res}{single integer, resolution of png image to create (px/inch)}

\item{alpha}{single integer (0-255), transparency of error bars (default 50)}
}
\description{
Creates a barplot or PNG file for each selected row
of each matrix provided, where each horizontal bar is
one column in that matrix.
}
\examples{
pairs <- list("2013" = pairLookup13,
    "2014" = pairLookup14, "2015" = pairLookup15)
allbarplots(lookups = pairs,
    lists = khador)
}

