% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/doRound.R
\name{doRound}
\alias{doRound}
\title{Get Results for a Round}
\usage{
doRound(data, round = 7, results = NULL, ratings, pairlookup,
  pairs = NULL)
}
\arguments{
\item{data}{data frame with columns:\itemize{
    \item Team
    \item Player
    \item List
}}

\item{round}{single integer}

\item{results}{NULL or data frame with a row named for each team
and with columns:\itemize{
    \item Total.Games.Won numeric total number of games won
    \item Matches.Won numeric total number of matches won
    \item Opponent.Round.X one or more columns naming the opponent in each round
}}

\item{ratings}{\link[PlayerRatings]{rating} object}

\item{pairlookup}{matrix of List pairings}

\item{pairs}{character vector of unique teams}
}
\value{
data frame with rownames Team, and columns Total.Games.Won, Matches.Won,
and Opponent.Round.X, where X is the value of round
}
\description{
Given Players in Teams with Lists, with ratings and taking into
account list pairs, calculate results for a given round.
Take into account results for previous rounds if provided.
}
\examples{
rat <- steph(x = wtc2014[, c("round", "player1", "player2", "TP")])
players2014 <- unique(c(wtc2014$player1, wtc2014$player2))
players2015 <- unique(wtc2015_players$Player)
rat$ratings <- rbind(rat$ratings,
    data.frame(Player = players2015[!players2015 \%in\% players2014],
        Rating = 2200, Deviation = 300,
        Games = 0, Win = 0, Draw = 0, Loss = 0, Lag = 0))
doRound(data = wtc2015_players, results = NULL,
    ratings = rat, pairlookup = initializeLookup(data = wtc2015_players$List))
}

