% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/updateLookup.R
\name{getPairs}
\alias{getPairs}
\title{Unique combinations paired in two records.}
\usage{
getPairs(data)
}
\arguments{
\item{data}{matrix, list or data.frame with two columns}
}
\value{
character matrix with two columns
}
\description{
Pairings of X vs X are not included.
}
\examples{
dat1 <- data.frame(list1 = c("a", "b", "b"), list2 = c("b", "b", "a"))
getPairs(data = dat1)
dat2 <- data.frame(list1 = c("A", "A", "B"), list2 = c("C", "B", "D"))
getPairs(data = dat2)
}
