% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getStat.R
\name{getStat}
\alias{getStat}
\title{Get statistic from difference between predicted value and scorefrac}
\usage{
getStat(data, status = NULL, pairlookup, compare = "scorefrac",
  round = "round", player1 = "player1", player2 = "player2",
  result = "result", list1 = "list1", list2 = "list2")
}
\arguments{
\item{data}{data.frame with columns round, 
player1, player2, result, list1, list2, compare}

\item{status}{data frame containing prior ratings (from class rating)}

\item{pairlookup}{square matrix with rownames in list1 and colnames in list2.}

\item{compare}{single character name of comparison column (default "scorefrac")}

\item{round}{single character name of round column (default "round")}

\item{player1}{single character name of player one column (default "player1")}

\item{player2}{single character name of player two column (default "player2")}

\item{result}{single character name of result column (default "result")}

\item{list1}{single character name of player one's list column (default "list1")}

\item{list2}{single character name of player two's list column (default "list2")}
}
\description{
Use additional information about closeness of game 
to find goodness of prediction.
}
\examples{
pl <- matrix(data = c(0, 30, -30, 0), nrow = 2, 
    dimnames = list(c("a", "b"), c("a", "b")))
dat <- data.frame(round = rep(1:6, each = 2),
    player1 = rep(c("A", "B"), times = 2), 
    player2 = rep(c("B", "A"), times = 2),
    result = c(0, 1, 1, 1, 0, 1, 0, 0, 0, 1, 1, 0), 
    list1 = c("a", "b", "b", "a"), 
    list2 = c("b", "a", "a", "b"), 
    scorefrac = c(0.1, 0.8, 0.7, 0.8, 0.3, 0.9, 0.1, 0.2, 0, 0.9, 0.8, 0.1), 
    stringsAsFactors = FALSE)
getStat(data = dat, pairlookup = pl, result = "result")
}
