% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/harmonizeCasters.R
\name{harmonizeCasters}
\alias{harmonizeCasters}
\title{Convert Warcaster and Warlock names to standard format}
\usage{
harmonizeCasters(x)
}
\arguments{
\item{x}{single character vector of caster names}
}
\value{
single character vector of harmonized caster names
}
\description{
Inserts space between name and number. 
In addition adds 1 if no number, and manually updates
some common alternative names.
}
\examples{
harmonizeCasters(c("Vlad", "Vlad1", "Vald 1"))
}
