% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/updateLookup.R
\name{initializeLookup}
\alias{initializeLookup}
\title{Create a square matrix for pairing strength scores}
\usage{
initializeLookup(data, list1 = "list1", list2 = "list2")
}
\arguments{
\item{data}{data.frame with columns round,
player1, player2, result, list1, list2, compare}

\item{list1}{single character name of player one's list column (default "list1")}

\item{list2}{single character name of player two's list column (default "list2")}
}
\description{
Generate an empty pairlookup matrix with columns named in
either list1 or list2.
}
\examples{
dat <- data.frame(list1 = c("A", "A", "B"), list2 = c("C", "B", "D"))
initializeLookup(data = dat)
}

