% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/scoreSequence.R
\docType{data}
\name{leaderboard15}
\alias{leaderboard15}
\title{World Team Championships 2015 Team Leaderboard}
\format{Data frame with columns:\itemize{
    \item  Rank integer noting team position
    \item  Team character name of team
    \item  Rounds.won integer number of rounds won
    \item  Team.SoS integer strength of schedule
    \item  Games.won integer number of games won
    \item  Total.CP integer control points scored
    \item  Total.VP integer army points scored
}}
\description{
Team results for WTC 2015.
}
\examples{
hist(leaderboard15$Total.CP, col = "#99334455")
}
\keyword{datasets}

