% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/makeMatchup.R
\name{makeAllMatchups}
\alias{makeAllMatchups}
\title{Pair All Teams}
\usage{
makeAllMatchups(results)
}
\arguments{
\item{results}{NULL or data frame with a row named for each team
and with columns:\itemize{
    \item Total.Games.Won numeric total number of games won
    \item Matches.Won numeric total number of matches won
    \item Opponent.Round.X one or more columns naming the opponent in each round
}}

\item{maxit}{single integer number of times to repeat}
}
\value{
character vector length number of rows results
}
\description{
Returns a vector where each team in an odd position plays the
team in the following even position.
Teams on the same Total Games Won are preferentially paired.
If there is an odd number of teams on a given score,
the team with the highest Matches Won in the next score down is paired.
}
\examples{
eg <- data.frame(Total.Games.Won = c(5, 4, 0, 1),
    Matches.Won = c(1, 1, 0, 0),
    Opponent.Round.1 = c("c", "d", "a", "b"),
    stringsAsFactors = FALSE)
rownames(eg) <- letters[1:4]
makeAllMatchups(results = eg)
}

