% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/makeMatchup.R
\name{makeMatchup}
\alias{makeMatchup}
\title{Pair Teams on the Same Score}
\usage{
makeMatchup(results, maxit = 10L)
}
\arguments{
\item{results}{NULL or data frame with a row named for each team
and with columns:\itemize{
    \item Total.Games.Won numeric total number of games won
    \item Matches.Won numeric total number of matches won
    \item Opponent.Round.X one or more columns naming the opponent in each round
}}

\item{maxit}{single integer number of times to repeat}
}
\value{
character vector length number of rows results
}
\description{
Returns a vector where each team in an odd position plays the
team in the following even position.
}
\examples{
eg <- data.frame(Total.Games.Won = c(5, 4, 4, 3),
    Matches.Won = c(1, 1, 1, 1),
    Opponent.Round.1 = c("e", "c", "b", "f"),
    Opponent.Round.2 = c("b", "a", "f", "e"),
    stringsAsFactors = FALSE)
rownames(eg) <- letters[1:4]
makeMatchup(results = eg)
}

