% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/playMatchup.R
\name{makePairing}
\alias{makePairing}
\title{Make Random Pairings}
\usage{
makePairing(team1players, team2players, team1lists, team2lists, round = 7)
}
\arguments{
\item{team1players}{length 5 or 10 character vector}

\item{team2players}{length 5 or 10 character vector}

\item{team1lists}{length 10 character vector}

\item{team2lists}{length 10 character vector}

\item{round}{single integer match round}
}
\value{
data frame with 5 rows and columns:\enumerate{
    \item round
    \item team1players
    \item team2players
    \item team1lists
    \item team2listss
}
}
\description{
Each player meets randomly.
Each player selects their list randomly.
}
\examples{
makePairing(team1players = letters[1:5],
    team2players = letters[11:15],
    team1lists = LETTERS[1:10],
    team2lists = LETTERS[11:20])
}

