% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/playMatchup.R
\name{playMatchup}
\alias{playMatchup}
\title{Simulate the result of a Match}
\usage{
playMatchup(team1, team2, data, ratings, pairlookup, round = 7)
}
\arguments{
\item{team1}{single character name of Team in data}

\item{team2}{single character name of Team in data}

\item{data}{data frame with columns:\itemize{
    \item Team
    \item Player
    \item List
}}

\item{ratings}{\link[PlayerRatings]{rating} object}

\item{pairlookup}{matrix with gamma correction for
matchup pairing with column and row names in List}

\item{round}{single integer}
}
\value{
length 5 numeric
}
\description{
Get the results of 5 games between 10 players in each of two teams.
}
\examples{
set.seed(43626)
playMatchup(team1 = "China", team2 = "Canada Goose",
    data = dwtc2015, ratings = rating2015, pairlookup = pairLookup)
## [1] 0 1 1 0 1
}

