% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/scoreSequence.R
\name{scoreSequence}
\alias{scoreSequence}
\title{Create Score for Predicted Sequence}
\usage{
scoreSequence(guess, result)
}
\arguments{
\item{guess}{character vector, 1 or more elements of result
starting from position 1}

\item{result}{character vector, comparison values}
}
\value{
single numeric
}
\description{
The score is the mean distance a guess position is from
the actual result. The score is sum of the absolute distance between
the predictions and the actual results, divided by the number
of guesses.
}
\details{
To allow people to guess country rather than team, there
is also handling for partial matches of teams so that "Team USA" matches
"Team USA Stars" and "Team USA Stripes". This will be scored as one
guess which has the average distance to each match.
}
\examples{
scoreSequence(guess = c("c", "b", "a"), result = letters) # 4/3 (3)
scoreSequence(guess = letters[1:3], result = letters) # 0 (3)
scoreSequence(guess = c("a", "a", "a"), result = letters) # 1 (3)
scoreSequence(guess = c("aa", "ab", "a"), result = paste0(letters, "b")) # 29/3 (3)
scoreSequence(guess = "Australia", result = leaderboard15$Team) # 2 (1)
}

