% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{showSimilarNames}
\alias{showSimilarNames}
\title{Print names that are similar}
\usage{
showSimilarNames(s, dist = 4L)
}
\arguments{
\item{s}{character vector}

\item{dist}{integer specifying how close records should be}
}
\description{
use \code{adist} to find matrices of distances
between names. First letter is not usually incorrect, 
so search alphabetically.
}
\examples{
showSimilarNames(c("Minnie", "Millie", "Ernestine"))
}
