% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/playMatchup.R
\name{splitPairs}
\alias{splitPairs}
\title{Split a vector}
\usage{
splitPairs(n = 10L)
}
\arguments{
\item{n}{single even integer}
}
\value{
logical vector length n
}
\description{
Splits sequential pairs in a vector randomly into two groups.
}
\examples{
splitPairs(n = 2)
splitPairs(n = 4)
}

