% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/allbarplots.R
\name{tbarplot}
\alias{tbarplot}
\title{Horizontal Barplot}
\usage{
tbarplot(x, list1, list2, pmain = "", penalty = 10, col = "#FF000022",
  xlim, bands = c(1, 3, 10), ...)
}
\arguments{
\item{x}{matrix, with symmetric row names and column names}

\item{list1}{single character, names from x}

\item{list2}{character vector, names from x}

\item{pmain}{single character main title}

\item{penalty}{single numeric, step size in x}

\item{col}{single colour for bars}

\item{xlim}{length 2 numeric, range of x axis}

\item{bands}{numeric vector, number of band multiples}
}
\value{
numeric vector of band positions invisibly.
}
\description{
Horizontal Barplot
}
\examples{
nms <- c("Butcher3", "Caine2", "Issyria", "Saeryn")
pl <- matrix(c(0:3, 1, 0:3, 2, 0:3, 1, 0),
    nrow = 4, ncol = 4, dimnames = list(nms, nms))
attr(pl, "n") <- pl
tbarplot(x = pl, list1 = "Butcher3", list2 = nms[-1])
}

