% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/updateLookup.R
\name{updateLookup}
\alias{updateLookup}
\title{Update pair matrix using data.}
\usage{
updateLookup(data, status = NULL, pairlookup = NULL, pairs = NULL,
  penalty = 10, seed = NULL, compare = "scorefrac", round = "round",
  player1 = "player1", player2 = "player2", result = "result",
  list1 = "list1", list2 = "list2", inflate = 1.1)
}
\arguments{
\item{data}{data.frame with columns round, 
player1, player2, result, list1, list2, compare}

\item{status}{data frame containing prior ratings (from class rating)}

\item{pairlookup}{square matrix with rownames in list1 and colnames in list2.}

\item{pairs}{matrix with two colums
where each row is a pair of lists to optimize}

\item{penalty, }{single numeric, magnitude of maximum penalty 
available to pairs (default 30)}

\item{seed}{single numeric, reproducible randomization}

\item{compare}{single character name of comparison column (default "scorefrac")}

\item{round}{single character name of round column (default "round")}

\item{player1}{single character name of player one column (default "player1")}

\item{player2}{single character name of player two column (default "player2")}

\item{result}{single character name of result column (default "result")}

\item{list1}{single character name of player one's list column (default "list1")}

\item{list2}{single character name of player two's list column (default "list2")}

\item{inflate}{single numeric in interval (1, 2), allow optimizer to search 
further when there are more results for a given pair.}
}
\value{
matrix with rownames in list1 and colnames in list2.
}
\description{
Optimization is restricted to +/- penalty * number of records with pairing.
Pairs are shuffled before optimization.
}
\examples{
dat <- data.frame(round = rep(1:6, each = 2),
    player1 = rep(c("A", "B"), times = 2), 
    player2 = rep(c("B", "A"), times = 2),
    result = c(0, 1, 1, 1, 0, 1, 0, 0, 0, 1, 1, 0), 
    list1 = c("a", "b", "b", "a"), 
    list2 = c("b", "a", "a", "b"), 
    scorefrac = c(0.1, 0.8, 0.7, 0.8, 0.3, 0.9, 0.1, 0.2, 0, 0.9, 0.8, 0.1), 
    stringsAsFactors = FALSE)
updateLookup(data = dat)
}
