% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/WTCTools-package.R
\docType{data}
\name{wtc2014}
\alias{wtc2014}
\title{World Team Championships 2014 Results}
\format{Data frame with columns:\itemize{
    \item game_id integer unique identifier for every match
    \item round integer for rounds 1 to 6
    \item TP integer 1 (players are ordered so that player 1 is always the winner)
    \item victory_condition character description of how match ended
    \item player1 character name of player
    \item team1 character team name
    \item list1 character warcaster/warlock
    \item faction1 character faction
    \item CP1 integer control points scored
    \item AP1 integer army points scored
    \item player2 character name of player
    \item team2 character team name
    \item list2 character warcaster/warlock
    \item faction2 character faction
    \item CP2 integer control points scored
    \item AP2 integer army points scored
}}
\description{
Downloaded from http://wmh-wtc.com/query.php
and converted to wide format so that every match is represented 
by one record.
}
\examples{
hist(wtc2014$AP1, col = "#99334455", 
    xlim = range(c(wtc2014$AP1, wtc2014$AP2)), ylim = c(0, 150), 
    xlab = "Army Points", main = "")
hist(wtc2014$AP2, col = "#33887755", add = TRUE)
legend("topright", legend = c("Player 1", "Player 2"), 
    pch = 22, pt.bg = c("#99334455", "#33887755"))
}
\keyword{datasets}
