% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/WTCTools-package.R
\docType{data}
\name{wtc2015_players}
\alias{wtc2015_players}
\title{World Team Championships 2015 Players}
\format{Data frame with columns:\itemize{
    \item Team character team name
    \item Faction character faction
    \item Player character name of player
    \item List character warcaster/warlock
    \item Objective character name of ojective selected
    \item NModels integer number of entries in list
}}
\description{
Downloaded from http://www.discountgamesinc.com/tournaments/
}
\examples{
table(wtc2015_players$Objective)
par(las = 2, mar = c(10, 4, 1, 1))
barplot(table(wtc2015_players$Faction))
}
\keyword{datasets}

