% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/WTCTools-package.R
\docType{package}
\name{WTCTools-package}
\alias{WTCTools-package}
\title{Warmachine World Team Championships Player and List Rankings}
\description{
Some utilities for extracting warcaster/warlock pairings
from the WTC matchup results 2014.
}
\details{
WARMACHINE is a registered trademark of Privateer Press Inc.
}
\author{
Lacerto
}

