% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/allbarplots.R
\name{allbarplots}
\alias{allbarplots}
\title{Create Barplots for all Rows in a Matrix}
\usage{
allbarplots(lookups, lists, nmin = 3L, fileroot = NULL, col = "red",
  height = 800, width = 800, res = 100, alpha = 50, ...)
}
\arguments{
\item{lookups}{matrix or named list of matrices, 
each with symmetric row names and column names}

\item{lists}{character vector names from lookups}

\item{nmin}{single integer, minimum number of games for bar to be plotted (default 3L)}

\item{fileroot}{single character label to prepend to start of filenames,
perhaps including path, which will be constructed 
"<fileroot><listname><lookupname>.png".
If NULL, plots will be created in the windows device.}

\item{col}{single colour for bars}

\item{height}{single integer, height of png image to create (px)}

\item{width}{single integer, width of png image to create (px)}

\item{res}{single integer, resolution of png image to create (px/inch)}

\item{alpha}{single integer (0-255), transparency of error bars (default 50)}
}
\description{
Creates a barplot or PNG file for each selected row 
of each matrix provided, where each horizontal bar is 
one column in that matrix.
}
\examples{
wtc2014 <- na.omit(wtc2014)
wtc2014$scorefracCP <- wtc2014$CP1 / (wtc2014$CP1 + wtc2014$CP2)
## if player 2 scored 0
wtc2014$scorefracCP[wtc2014$CP2 == 0] <- 0.5 + wtc2014$CP1[wtc2014$CP2 == 0] / (5 * 2)
## if player 1 scored 0
wtc2014$scorefracCP[wtc2014$CP1 == 0] <- 0.5 - wtc2014$CP2[wtc2014$CP1 == 0] / (5 * 2)
## fractional army points (approximately)
wtc2014$AP1 <- wtc2014$AP1 / 61
wtc2014$AP2 <- wtc2014$AP2 / 61
## ratio of score
wtc2014$scorefracAP <- wtc2014$AP1 / (wtc2014$AP1 + wtc2014$AP2)
## if player 2 scored 0
wtc2014$scorefracAP[wtc2014$AP2 == 0] <- 0.5 + wtc2014$AP1[wtc2014$AP2 == 0] / 2
## if player 1 scored 0
wtc2014$scorefracAP[wtc2014$AP1 == 0] <- 0.5 - wtc2014$AP2[wtc2014$AP1 == 0] / 2
## combine the scores
wtc2014$scorefrac <- 0.5 * wtc2014$scorefracCP + 0.5 * wtc2014$scorefracAP
wtc2014$scorefrac[wtc2014$scorefrac > 1] <- 1
\dontrun{
pairLookup14 <- updateLookup(data = wtc2014, 
    pairlookup = initializeLookup(data = unique(c(wtc2014$list1, wtc2014$list2))), 
    penalty = 10,
    result = "TP")
khador <- c("Butcher", "Butcher2", "Butcher3", "Harkevich", "Irusk", "Irusk2", 
    "Karchev", "Old Witch", "Sorscha", "Sorscha2", "Strakhov", "Vlad", "Vlad2", "Vlad3", 
    "Zerkova2")
allbarplots(lookups = pairLookup14, 
    lists = khador)
}
}
