% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/updateLookup.R
\name{initializeLookup}
\alias{initializeLookup}
\title{Create a square matrix for pairing strength scores}
\usage{
initializeLookup(data)
}
\arguments{
\item{data}{Character vector or list of character vectors}
}
\description{
Create a blank named matrix with attribute "n",
which is an identical blank matrix.
}
\examples{
dat <- data.frame(list1 = c("A", "A", "B"), 
    list2 = c("C", "B", "D"), res = 1:3)
initializeLookup(data = dat[c("list1", "list2")])
}
