% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/getMatrixVal.R
\name{setMatrixVal}
\alias{setMatrixVal}
\title{Set a value in a Matrix}
\usage{
setMatrixVal(list1, list2, x, val)
}
\arguments{
\item{list1}{name}

\item{list2}{name}

\item{x}{matrix}

\item{val}{single numeric}
}
\value{
matrix
}
\description{
Updater function for named matrix
}
\examples{
m1 <- matrix(0, nrow = 3, ncol = 3, dimnames = list(letters[1:3], letters[1:3])) / 10
setMatrixVal(list1 = "b", list2 = "a", x = m1, val = -2)
setMatrixVal(list1 = c("a", "b", "c"), list2 = c("b", "c", "a"), x = m1, val = 1:3)
}

