% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/allbarplots.R
\name{tbarplot}
\alias{tbarplot}
\title{Horizontal Barplot}
\usage{
tbarplot(x, list1, list2, pmain = "", penalty = 10, nmin = 3L,
  col = "#FF000022", xlim, bands = c(1, 3, 10), mar = c(2, 7, 2, 1),
  las = 1, ...)
}
\arguments{
\item{x}{matrix, with symmetric row names and column names}

\item{list1}{single character, names from x}

\item{list2}{character vector, names from x}

\item{pmain}{single character main title}

\item{penalty}{single numeric, step size in x (default 10)}

\item{nmin}{single integer, minimum number of games for bar to be plotted (default 3L)}

\item{col}{single colour for bars}

\item{xlim}{length 2 numeric, range of x axis}

\item{bands}{numeric vector, number of band multiples (default c(1, 3, 10))}
}
\value{
numeric vector of band positions invisibly.
}
\description{
Create barplots for a row of a pairlookup matrix.
}
\examples{
pl14 <- structure(c(4.80673409037624, 21.3865173349445, -5.17793802916065, 
        45.9138959155639, 14.5926699048978, -21.3865173349445, -33.4375895685107), 
    .Dim = c(1L, 7L), 
    .Dimnames = list("Butcher3", c("Caine2", "Deneghra", "Haley", "Haley2", 
        "Issyria", "Kreoss3", "Saeryn")))
attr(x = pl14, which = "n") <- structure(c(3, 2, 2, 4, 3, 2, 3), 
    .Dim = c(1L, 7L), 
    .Dimnames = list("Butcher3", c("Caine2", "Deneghra", "Haley", "Haley2", 
        "Issyria", "Kreoss3", "Saeryn")))
nm <- colnames(pl14)[order(pl14["Butcher3", ])]
tbarplot(x = pl14, list1 = "Butcher3", list2 = nm)
}
